+++ sample/PractiseBlending30.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 
 static double deltaTime;
-static float lastFrame;
+static double lastFrame;
 static double lastX = SMALL_SCREEN_WIDTH / 2, lastY = SMALL_SCREEN_HEIGHT / 2;
 static bool isMouseFirstIn = true;
