+++ sample/TestPractiseLight27.cpp
@@ -1,7 +1,7 @@
 #include "../include/TestPractiseLight27.h"
 
 static double deltaTime;
-static float lastFrame;
+static double lastFrame;
 static double lastX = SMALL_SCREEN_WIDTH / 2, lastY = SMALL_SCREEN_HEIGHT / 2;
 static bool isMouseFirstIn = true;
 
@@ -157,7 +157,7 @@ static void mouse_move_callback(GLFWwindow* window, double posX, double posY) {
 }
 
 static void mouse_scroll_callback(GLFWwindow* window, double offsetX, double offsetY) {
-	camera.ProcessMouseScroll(offsetY);
+	camera.ProcessMouseScroll((float)offsetY);
 }
