+++ sample/testPractiseLight23.cpp
@@ -136,7 +136,7 @@ int practiseLight23(const char * projectDir) {
     glfwSetCursorPosCallback(window, mouse_move_callback);
 
     while (!glfwWindowShouldClose(window)) {
-        double curFrame = glfwGetTime();
+        float curFrame = (float)glfwGetTime();
         deltaTime = curFrame - lastFrame;
         lastFrame = curFrame;
 
@@ -290,6 +290,6 @@ void mouse_move_callback(GLFWwindow* window, double posX, double posY) {
 }
 
 void mouse_scroll_callback(GLFWwindow* window, double offsetX, double offsetY) {
-    camera.ProcessMouseScroll(offsetY);
+    camera.ProcessMouseScroll((float)offsetY);
 }
